Golden change for float extra_t1d (#252)
One of the goldens has integer extra_t1d which are going to cause
problems with the extra_t2d branch. This is the simplest possible change
to fix the golden.

The actual code changes are going to get nuked by extra_t2d, but, the
golden ends up correct.

Before
<img width="881" alt="Screen Shot 2024-05-23 at 6 53 44 PM"
src="https://github.com/baker-laboratory/rf_diffusion/assets/21205950/456ed413-66bc-4114-bdcc-11986eb9e1b9">

After:
<img width="854" alt="image"
src="https://github.com/baker-laboratory/rf_diffusion/assets/21205950/cd52bff2-328b-4713-a135-434b81e60883">

+++ rf_diffusion/features.py
@@ -8,12 +8,12 @@ from rf_diffusion import sasa
 
 def get_extra_t1d(indep, featurizer_names, **kwargs):
     if not featurizer_names:
-        return torch.zeros((indep.length(),0))
+        return torch.zeros((indep.length(),0)).float()
     t1d = []
     for name in featurizer_names:
         feats_1d = featurizers[name](indep, kwargs[name], **kwargs)
         t1d.append(feats_1d)
-    return torch.cat(t1d, dim=-1)
+    return torch.cat(t1d, dim=-1).float()
 
 def one_hot_bucket(x: torch.Tensor, boundaries: torch.Tensor):
     '''
@@ -157,11 +157,11 @@ inference_featurizers = {
 
 def get_extra_t1d_inference(indep, featurizer_names, params_train, params_inference, **kwargs):
     if not featurizer_names:
-        return torch.zeros((indep.length(),0))
+        return torch.zeros((indep.length(),0)).float()
     t1d = []
     for name in featurizer_names:
         assert name in params_train
         assert name in params_inference
         feats_1d = inference_featurizers[name](indep, params_train[name], params_inference[name], **kwargs)
         t1d.append(feats_1d)
-    return torch.cat(t1d, dim=-1)
+    return torch.cat(t1d, dim=-1).float()
